Updated some decimation and added in the default variable
This is a really cool project and worked great, but I was putting the -z option last and then it wouldn't work. So I just added an example in that shows an example usage and then I added a initial declaration to the top of the script so that it has a default option as described in -h.